Move location fields out of user data
This is an alternative approach to https://github.com/dimagi/commcare-hq/pull/35292
I move location data out of user data entirely, injecting it at the
last minute.  We already did this for web users as of
https://github.com/dimagi/commcare-hq/pull/34401, but there, the fields
were _always_ present, whereas for mobile workers, they're only present
if not null.  This commit brings web users in line with the behavior for
mobile workers, for better parity