Add use_all_objects option
This option is useful when the default object manager of a model class
does not behave as needed when dumping objects. For example, if deleted
objects need to be included in the dump due to other models referencing
those objects, this ensures a default object manager that ignores
deleted objects isn't used.