Cache file upload original names
When deleting a repeat group we need to rebuild all the following groups
because the IDs are reassigned by formplayer. Unfortunatly FP does not
keep track of the original filename automatically and only sends the
uuid based generated filename for the upload in the delete-repeat
response. To be able to still display the original filename we store it
in a map with the generated name as the key. That way we can figure out
what it is when we rebuild that part of the UI.