Empty strings in recipient filter match unset prop
If a user field/property is created after the user and the user
has not been updated yet, it is not in the user data or user case
properties. In that case the conditional alert recipient filter
will fail if it uses said field. With this change a filter value
"" will match if a field does not exist yet in the user data or
user case properties.