Updated mocha/base.html to support webpack
Notably, in requirejs, tests do not include "base_main" js,
which includes a bunch of global UI-oriented code: analytics,
modals for accounting and for inactivity, etc. In webpack, these
are part of the "commcarehq" bundle, which can end up in the dependency
tree for tests - as an example, for hqwebapp, hqwebapps/js/widgets is a
dependency of tests, and it includes commcarehq. To deal with this, I'm
updating this template to not error out on this UI code. This is why the language script
is moved up (by moving it above webpack.html, it's now available to all of the
js in tests), and why the commcarehq-urls block is added (without this,
notifications_service, which is part of base_main, errors out because it depends
on querying the notifications url).