bug: correctly handle "None" vs "''" inputs.
There is a slight difference in how bulk upload and the API works. For bulk upload, the existing
user information is typically included. So if the field is required, then the existing value is expected
to be included in the bulk upload. If the field is either empty ("''") or does not exist (None), validation
will failed as "the field is required" for bulk upload. However for the API, if it is a PUT and the field is not
included (None), then that should not fail the required validation.