Combined standard report logic and UCR logic
This combines the UCR version of getStandardHQReport, which the breaking change caused to no longer get called,
into the logic in standard_hq_report.js. It sets up a js_options dictionary to use in UCR, which was previously
using initial page data.

It also eliminates the old block for handling custom reports, which should be irrelevant now that all custom
reports use external js files and hqDefine, instead of ineline js and global variables.