Replace settings limits with SystemLimit
This commit did a lot. In addition to replacing the limit in settings
with SystemLimit, I removed the feature flag for an increased limit
since SystemLimit solves that, and moved the setting to enable the
device rate limiter to a feature flag for ease of use. I also updated
metrics to differentiate between an actual rate limited attempt vs a
would-be rate limited.