Added explicit dependencies to app_manager.js
Needed to do this to fix tests, which use webpack which means all dependencies need to have their dependencies
declared. Also updated HTML to ensure all app manager dependencies are loaded in the correct order
for the actual rendeering of the page, which does not use webpack. (The non app manager dependencies are all
global-ish dependencies that are definitely included before any of the app manager ones.)