Local cache with 5 min timeout instead of clearing
In the worst case, it'll take 5 minutes for a new limit to take effect,
but we don't have to worry about clearing cache.

This also uses local memory instead of redis, and because the underlying
db query isn't too expensive, it should be more efficient to rely on a
shorter cache life in local memory than a longer cache life in redis.