Revert to simpler cache
This references settings.UNIT_TESTING in
production code to skip caching, but this is a less fragile approach
than the solution I had to remove the quickcache decorator by blindly
unwrapping the function, and is used throughout our code.

This uses redis in addition to local memory caching, but that is
how it is typically done in HQ and has worked well for us, and results
in a simpler decorat that is easier for future readers to understand.

This stores redundant data in the cache, but avoids hitting the db again
which is what we want.