Remove select all pending and cancelled buttons
Again, this can already be done by filtering the report to the specific
state you want to take action on. These buttons are also a pain to
manage as repeat record states evolve. For example, it isn't clear to a
user what a pending record is, since that could be Failed or Pending.