Require superuser access to resend records
In addition to adding confusion to this page (resend vs requeue),
resending gives users the ability to "skip the line" for data
forwarding. Practically, requeueing gives a typical user everything they
need to retry records. In a pinch, it is nice to skip the line and
resend records immediately, but this should only be given to superusers.