Match none only if value is None
It is likely better to have the caller decide what the behavior should
be in this scenario, but the conservative approach is to match none if a
value of None is specified. For the sake of resolving a very minor bug
that is behind a deprecated feature flag, this seems fine.