Disable deleting system properties
This function is for disconnecting a CommCare case from an associated
OpenMRS Patient (for example, if that Patient instance was merged with
another Patient instance). The case property that is passed to the
function is the one configured to store the ID of the OpenMRS patient.
This change is to prevent a misconfigured integration from deleting
system case properties. See unit tests for examples.