Fix bug in LocationManager.filter_by_user_input
Always and only filter the root query by domain, even when it has no
user input. Lower query levels need not be filtered by domain since
those levels are constrained by parent_id, where the parent should
always be in the same domain as its children.

Fixes an issue with no user input on the first level, which caused

  query = self.get_queryset_descendants(query)

to raise

  AttributeError: 'LocationManager' object has no attribute 'id'

because `query` was `self` (a LocationManager object).