Pull location types stuff into its own page
It's conceptually separate, and it's not part of the locations API -
having it here makes it look like functionality missing from the v2, but
really it's something standalone.  I'd like to remove it entirely
anyways, but whatever.

I could've split this out to preserve history like I did with the main
API pages, but I got lazy