Use record_id parameter to determine code path
State was not reliable as it didn't indicate if the entire table or
specific record ids were being selected. This change makes it so that we
look at the record_id parameter to determine if we should process an
entire queryset of repeat records based on report filters, or just a set
list of record ids passed up by the user.