Add "dev" dependency group
The "sso" group was moved to "default-groups" to allow it to be excluded
with `--no-group=sso` when running commands like `uv sync`. The
alternative of adding it to various groups with `{include-group = "sso"}`
made it impossible to install those groups without "sso" dependencies.

"docs" requirements were not included in the "dev" group to keep it
light. They can be easily installed with `uv sync --group=docs`.

Copied test-requirements.in to dependency group in pyproject.toml.
Then ran the following commands:

    uv pip install -r requirements/dev-requirements.txt

    # lock versions from output of previous command
    # adjusted as necessary to resolve inconsistencies
    uv lock \
      --upgrade-package='asttokens==2.4.1' \
      --upgrade-package='beaker==1.11.0' \
      --upgrade-package='beautifulsoup4==4.12.2' \
      --upgrade-package='build==1.2.2.post1' \
      --upgrade-package='click==8.1.3' \
      --upgrade-package='colorama==0.4.6' \
      --upgrade-package='gnureadline==8.2.13' \
      --upgrade-package='oic==1.7.0' \
      --upgrade-package='psutil==7.0.0' \
      --upgrade-package='pyproject-hooks==1.2.0' \
      --upgrade-package='snowballstemmer==2.1.0' \
      --upgrade-package='soupsieve==2.5' \
      --upgrade-package='wheel==0.45.1' \
      --upgrade-package='yapf==0.43.0'

    rm requirements/dev-requirements.*