Use state.name instead of label
We no longer display the state field in the UI, and therefore the
non-presentable name can be used (e.g., "InvalidPayload" instead of
"Invalid Payload"). When translating state back on bulk actions
requests, we use state.name to identify the state, and therefore
require state.name be used in the first place.