remove get_change_counts After some trials on staging and locally, it’s clear that the aggregate query adds too much overhead for the table. The check is now more detrimental than beneficial, since this check is run both on cell-level interactions and a full table refresh.
Previously we relied on get_change_counts to ensure the “number of records with an edit history” shown to the user stays up-to-date. However, since we no longer need to show this value to the user, the benefit of this aggregate query is minimal.

- num_changes replaces the usage of num_edited_records, as it is already being used in the UI and is an equally valid check to determine whether to show the user the edit bar.

Due to the more expensive nature of the aggregate query, checking wether a record has gone over 20 “chained” changes adds more detriment than benefit. We already have a session-wide check for overall changes. Thus, the showApplyWarning notice isn't beneficial due to the costly nature of the check needed to show it. Therefore, we can just get rid of it.