show edit bar based on template context Reconcile both storing this value in an Alpine.store and in the context.
We do not need to display a live number of num_changes anywhere, and the actions that remove changes always involve a full table referesh, so this context is always kept up-to-date.

For context, cell-only interactions never remove changes from the session, they only add to it. Hitting the X adds a RESET change. Only the “clear” and “undo” table-level interactions can remove changes.