Give DataSourceRepeater a real payload model
DataSourceRepeater currently uses dataclass DataSourceUpdateLog as a kind of
payload model, which it creates on the fly. This is different from all
other Repeater classes. DataSourceUpdate will bring DataSourceRepeater
in line with other Repeater class behavior, and will also allow us to
update the payload with new data as rows get added to a data source,
just as CaseRepeater payloads are updated as form submissions modify a case.