Remove ko-template class from any descendant
Knockout bindings cannot be applied to nested elements, so it should be
safe to remove the ko-template class from any descendant of the element
we are applying bindings too.

This enables us to add the ko-template class to elements within our
template even if applying bindings to parent elements that are defined
in an inherited template.