Add reassign_cases view for non-bulk workflow
I don't like how this view is far away from the related logic in the
data_interfaces app, but copy_cases is its sister and already here.

A 'reassign_cases' url already exists in the geospatial (microplanning)
app, so a different name had to be chosen for the new url, and the
'copy_cases' url was renamed to 'copy_cases_action' for consistency.
URL references in templates will be updated in a subsquent commit.