Build test docker image with Github Actions
A Github Action provides more control over the image than is available
with Docker Hub build automation. Specifically, it allows the Python
version to be embedded in the image name.

As of this writing, the dimagi/commcare-hq repository cannot be selected
in the Docker Hub build settings, likely due to Github access control
changes since the build was originally configured. This means the
commcarehq_base image cannot currently be reconfigured or updated on
Docker Hub. The plan is to abandon that image with Python 3.9 and begin
using images generated by this Action for Python 3.13 and onward.

Encoding the build configuration as a source code artifact and Github
Action allows us to more reliably maintain it over time and more easily
monitor build status in association with a hypothetical PR that breaks
the build.