Revert support for null owner in scheduled report
If the owner is None, the scheduled report is deleted and therefore
should never hit this code. There is an edge case where another user
hits "Send Now" on the scheduled report, but given the hard deleted
owner is already such an edge case, I don't think it is worth attempting
to address this edge case right now.