Setup venv before switching user
COPY seems to be done as root, and then RUN is run as the current user.
We were attempting to rm -rf /tmp-project as the cchq user, but failed
to do so due to lack of permissions. Presumably because the tmp-project
directory was initially created by root as part of the COPY directive.