Do not exclude migrating domains from UCR rebuilds
This was originally intended to exclude domains that were migrating to a
different HQ environment[1], but since then DomainMigrationProgress has
been adopted for many other kinds of migrations, most of which are not
relevant to UCR rebuilds.

No DomainMigrationProgress records exist on prod pertaining to migrating
to a different environment, and set_migration_started(...) is not used
to start that type of migration anywhere in code, so it is unclear how
this model was used for that purpose historically. A better
implementation would have checked for migrations with a specific slug,
but removal seems justified since that slug is unknown.

[1] https://github.com/dimagi/commcare-hq/pull/28532