Upgrade django-tastypie v0.14.6 -> v0.15.1
For compatibility with Django 5: django.utils.datetime_safe is removed.
The compatibility fix in tastypie is strange because it monkey patches
Django rather than adjusting and removing the deprecated usage. Also,
it still accesses the deprecated code path on django<5, so the warning
exception cannot be removed until after django is upgraded.

https://github.com/django-tastypie/django-tastypie/blob/03c4746f0a0f88628be5264bc8d4a19a0529b2f4/tastypie/compat.py#L24