Stop skipping fixtures for BHA
I have a theory that a race condition meant that disabling that feature
flag didn't actually take effect in time for the restores.  This would
explain intermittent reports of a new fixture not appearing, despite
this specific accommodation.  Feature flag lookups are cached heavily in
a couple different ways, and they should be correctly invalidated, but
it's conceivable that the invalidation took longer than it took a
restore to start for some of these users.

Clearing user data does consistently resolve the issue, which is
consistent with the theory.  If this change is not sufficient to resolve
the issue, we can force a full sync for all users like so:

```
Clear and resync all users who've synced in the last two weeks:
$ manage.py prime_formplayer_restores --domains co-carecoordination --last-synced-hours 336 --clear-user-data
```