Unify user audit report and export_audit_data cmd
This makes the output of the export_audit_data management command and
the user audit report the same. This way, support can use the report as
a reference for what data would be included in the export if requested.

The report is still limited to a max number of records, so things like
sorting by date after combining access and navigation audit events
doesn't feel too terrible.

Lastly, the csv writer already converts values of None to empty strings,
so moving the "domain or ''" logic into shared code is safe.