Increase timeout significantly
There is a scenario where task A is queued, then task B. If task A takes
the entire timeout period to process, task B will effectively timeout
waiting to process. This increases the timeout to a value that execution
time shouldn't come anywhere close to, so as long as lots of tasks for a
domain are not queued at the same time, this should provide enough of a
cushion.