Remove PillowtopCheckpointReset handler
PillowtopCheckpointReset is raised when the checkpoint in the database
has an unexpected value while updating the checkpoint after processing
one or more changes. Therefore attempting to process the batch again
on reset would result in duplicate change processing followed by another
reset error when the checkpoint is updated by process_offset_chunk() in
the except block.