Catch RuntimeError when indexing design docs
We saw an influx of this error:

RuntimeError: Failed to verify design document _design/tests after 5
attempts

It is worth digging deeper on this, but for now it seems fine to gracefully
handle this failure so that other design docs are preindex whereas it
exists early right now.

If you want to see this sentry error in the future, just remove the
try/except.