Don't close connections every time
While this is ideal, we have code that currently calls this inside of a
transaction, which causes issues. Yes, we can change the calling code to
not do that, but it is also the case that this is a hard issue to debug
if encountered, so I wouldn't want to keep this code without a better
way to prevent devs from getting into similar situations.