Switch task to only touch SQLUserData
Users are big and saving them is slow - this operation should be much
faster and impact less.  It is a slight behavior change, in that the
`user.save()` signals won't be called.  I think that's advantageous, but
I'm open to counter arguments.

I don't love that _remove_unrecognized both modifies its arg _and_ has a
return value, but I'm not sure of a clean alternative