Handle race where TaskRecord is created by worker
Since the original apply_async call does not create a TaskRecord until
after calling super().apply_async, it is possible for a worker to pick
up the task and create a TaskRecord when retrying the task. To mitigate
this, we use update_or_create instead, making no change if the
subsequent call to apply_async won the race condition since that
information is more up to date anyway.