Remove last_modified field
Given the TaskRecord will be deleted upon completion, the only value the
last_modified field added was in the rare instance where a task was
retried with different args/kwargs, or upon retrying, ran into an error
queueing a message to the broker.

Given how rare either of these circumstances are, the existence of a
last_modified field is misleading in most cases, since a task could be
retried, but no changes are made, so last_modified wouldn't change. This
makes it a redundant field since created_at already holds this info.