Bypass SSRF WAF rule when registering oauth apps
Specifically when testing an oauth flow locally, it is very common to
have a local development environment running on local IPs, which means
you need to specify redirect URIs in the oauth application like
"localhost:8000". The AWS WAF sees this as a potential SSRF threat and
blocks it, but given how reasonable this is for registering oauth apps,
we should allow it.