Bypass SSRF WAF rule when registering oauth apps
Specifically when testing an oauth flow locally, it is very common to
have a local development environment running on local IPs, which means
you need to specify redirect URIs in the oauth application like
"localhost:8000". The AWS WAF sees this as a potential SSRF threat and
blocks it, but given how reasonable this is for registering oauth apps,
we should allow it.

Because this is a class based view, and we call view.as_view() in
urls.py, we need to apply the waf_allow directly to that function for it
work properly in list_waf_allow_patterns