Fix UUID regex and config matching
From https://github.com/dimagi/commcare-hq/security/code-scanning/462
The \b characters in here work if the expression is written as a regex
literal (surrounded by slashes), but not as a string. They are
unnecessary in this context though, so I just removed them.

From https://github.com/dimagi/commcare-hq/pull/35849
The impact of this bug was that the regex never returned a match, meaning
that menus with UUIDs in the path (child case menus) would not save the
selected set of columns unless the same UUID was selected first.