Create TaskRecord prior to calling apply_async
This avoids the race condition of a celery worker completing and
attempting to delete a TaskRecord before the producer has created it.

Additionally, the sent field is no longer needed since it is redundant
of the error column. And all durable tasks will be responsible for
creating their own task_id. The TaskRecord table has a default of
uuid.uuid4 for the task_id, which is the primary key as well.