Minor cleanup
I had trouble getting this to work properly from the django manage
machine when dealing with migrating brokers. Ultimatley, it is primarily
important that this can run successfully from celery machines since that
is where it is kicked off from, and it will use the config on the celery
machine.

Also added some docs, cleaned up an issue with _is_worker_up, and
removed the broker from logs since it contains sensitive info.